refactor(editor): move search concerns to a dedicated search service
avoid to have search concerns in dashboard page component
